third_party/blink/web_tests/platform/generic/fast/events/blur-focus-window-should-blur-focus-element-expected.txt

Blurring and focusing the window should blur and focus the focused element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisObjects[0] is divElement
PASS events[0] is "blur"
PASS targets[0] is divElement
PASS thisObjects[1] is window
PASS events[1] is "blur"
PASS targets[1] is window
PASS thisObjects[2] is window
PASS events[2] is "focus"
PASS targets[2] is window
PASS thisObjects[3] is divElement
PASS events[3] is "focus"
PASS targets[3] is divElement
PASS thisObjects[4] is divElement
PASS events[4] is "blur"
PASS targets[4] is divElement
PASS thisObjects[5] is window
PASS events[5] is "blur"
PASS targets[5] is window
PASS thisObjects[6] is iframe
PASS events[6] is "focus"
PASS targets[6] is iframe
PASS thisObjects[7] is innerDiv
PASS events[7] is "focus"
PASS targets[7] is innerDiv
PASS thisObjects[8] is innerDiv
PASS events[8] is "blur"
PASS targets[8] is innerDiv
PASS thisObjects[9] is iframe
PASS events[9] is "blur"
PASS targets[9] is iframe
PASS thisObjects[10] is window
PASS events[10] is "focus"
PASS targets[10] is window
PASS thisObjects[11] is divElement
PASS events[11] is "focus"
PASS targets[11] is divElement
PASS successfullyParsed is true

TEST COMPLETE
